requirements.txt
python-fasthtml
uvicorn

# Error detection dependencies
portkey-ai
python-dotenv
pyyaml
